FEAT(bulk_runner_rs_v2)!: Add a version 2 with a lot of refinement.
Bug:
bulk_runner_rs_v2/src/query_controller/query_engine.rs::cli_dispatch
Still contains an issue where join handle or join result is sometimes
dropped.

For the tool to be consistent this needs to be resolved.
We not be able to use a par_iter here due to how rayon works.